Move M3D handling into a satellite library
