+++ lib/m3d/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "m3d"
+version = "0.1.0"
+workspace = "../.."
+authors = ["Dzmitry Malyshau <kvarkus@gmail.com>"]
+
+[dependencies]
+byteorder = "1.0"
+log = "0.4"
+obj = { version = "0.9", optional = true }
+ron = { version = "0.4", optional = true }
+serde = "1.0"
+serde_derive = "1.0"
